requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data structures and persistence
dataclasses-json>=0.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# LLM & Embedding Providers
ollama>=0.1.0
httpx>=0.24.0
# openai>=1.0.0 (Optional)
# anthropic>=0.8.0 (Optional)

# Optional: Vector search
# faiss-cpu>=1.7.0

# Optional: Browser automation
# playwright>=1.40.0

# Development tools
black>=23.0.0
mypy>=1.0.0
ruff>=0.1.0